removed modpack page moved to modal
