+++ theseus_gui/src/components/ui/ExportModal.vue
@@ -1,7 +1,7 @@
 <script setup>
-import { Button, Checkbox, Modal, SendIcon, XIcon } from 'omorphia'
+import { Button, Checkbox, Modal, SendIcon, XIcon, PlusIcon } from 'omorphia'
 import { PackageIcon, VersionIcon } from '@/assets/icons'
-import { ref } from 'vue'
+import { ref, toRef } from 'vue'
 import { export_profile_mrpack, get_potential_override_folders } from '@/helpers/profile.js'
 import { open } from '@tauri-apps/api/dialog'
 import { handleError } from '@/store/notifications.js'
@@ -24,9 +24,11 @@ defineExpose({
 
 const exportModal = ref(null)
 const nameInput = ref(props.instance.metadata.name)
+const exportDescription = ref('')
 const versionInput = ref('1.0.0')
 const files = ref([])
 const folders = ref([])
+const showingFiles = ref(false)
 
 const themeStore = useTheming()
 
@@ -70,6 +72,31 @@ const initFiles = async () => {
     },
     value,
   ])
+
+  if (props.instance.export_cache) {
+    const export_cache = toRef(props.instance.export_cache).value
+    if (export_cache.name) {
+      nameInput.value = export_cache.name
+    }
+    if (export_cache.version) {
+      versionInput.value = export_cache.version
+    }
+    if (export_cache.description) {
+      exportDescription.value = export_cache.description
+    }
+    for (const file of export_cache.overrides) {
+      files.value.forEach((f) => {
+        if (f.path === file) f.selected = true
+      })
+      folders.value.forEach((args) => {
+        args[1].forEach((child) => {
+          if (child.path === file) {
+            child.selected = true
+          }
+        })
+      })
+    }
+  }
 }
 
 await initFiles()
@@ -93,7 +120,9 @@ const exportPack = async () => {
       props.instance.path,
       outputPath + `/${nameInput.value} ${versionInput.value}.mrpack`,
       filesToExport,
-      versionInput.value
+      versionInput.value,
+      exportDescription.value,
+      nameInput.value
     ).catch((err) => handleError(err))
     exportModal.value.hide()
   }
@@ -123,11 +152,34 @@ const exportPack = async () => {
           </Button>
         </div>
       </div>
+      <div class="adjacent-input">
+        <div class="labeled_input">
+          <p>Description</p>
+
+          <div class="textarea-wrapper">
+            <textarea
+              v-model="exportDescription"
+              placeholder="My super-special modpack is the best of them all!"
+            />
+          </div>
+        </div>
+      </div>
+
       <div class="table">
         <div class="table-head">
-          <div class="table-cell">Select files and folders to include in pack</div>
+          <div class="table-cell row-wise">
+            Select files and folders to include in pack
+            <Button
+              class="sleek-primary collapsed-button"
+              icon-only
+              @click="() => (showingFiles = !showingFiles)"
+            >
+              <PlusIcon v-if="!showingFiles" />
+              <XIcon v-else />
+            </Button>
+          </div>
         </div>
-        <div class="table-content">
+        <div v-if="showingFiles" class="table-content">
           <div v-for="[path, children] of folders" :key="path.name" class="table-row">
             <div class="table-cell file-entry">
               <div class="file-primary">
@@ -260,5 +312,23 @@ const exportPack = async () => {
   display: flex;
   gap: var(--gap-sm);
   align-items: center;
+}
+
+.row-wise {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.textarea-wrapper {
+  // margin-top: 1rem;
+  height: 12rem;
+  textarea {
+    max-height: 12rem;
+  }
+  .preview {
+    overflow-y: auto;
+  }
 }
 </style>
