+++ theseus_gui/src/pages/instance/index.js
@@ -2,6 +2,5 @@ import Index from './Index.vue'
 import Mods from './Mods.vue'
 import Options from './Options.vue'
 import Logs from './Logs.vue'
-import Modpack from './Modpack.vue'
 
-export { Index, Mods, Options, Logs, Modpack }
+export { Index, Mods, Options, Logs }
